# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/contacts": {
      "get": {
        "operationId": "ContactsController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "List of all contacts.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ContactResponseDto"
                  }
                }
              }
            }
          }
        },
        "summary": "Get all contacts",
        "tags": [
          "contacts"
        ]
      },
      "post": {
        "operationId": "ContactsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "Data to create a new contact.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateContactDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The contact has been created successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContactCreatedResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid data — owner is not exactly one of retailer/recipient, or the owner already has a primary contact."
          }
        },
        "summary": "Create a new contact",
        "tags": [
          "contacts"
        ]
      }
    },
    "/contacts/{id}": {
      "get": {
        "operationId": "ContactsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Contact ID",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Contact found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContactResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Contact not found."
          }
        },
        "summary": "Get a single contact",
        "tags": [
          "contacts"
        ]
      },
      "patch": {
        "operationId": "ContactsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID of the contact to update",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "description": "New data for the contact.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateContactDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Contact updated successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContactCreatedResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid data — owner is not exactly one of retailer/recipient, or the owner already has a primary contact."
          },
          "404": {
            "description": "Contact not found."
          }
        },
        "summary": "Update a contact",
        "tags": [
          "contacts"
        ]
      },
      "delete": {
        "operationId": "ContactsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Contact ID",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Contact deleted successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Contact not found."
          }
        },
        "summary": "Delete a contact by its ID",
        "tags": [
          "contacts"
        ]
      }
    },
...
    "/invalid-tokens": {
      "get": {
        "operationId": "InvalidTokensController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "List of all denylist entries.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/InvalidTokenResponseDto"
                  }
                }
              }
            }
          }
        },
        "summary": "Get all denylist entries",
        "tags": [
          "invalid-tokens"
        ]
      },
      "post": {
        "operationId": "InvalidTokensController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "Data to add a token to the denylist.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateInvalidTokenDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The token has been revoked successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvalidTokenCreatedResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid data — the jti is already on the denylist."
          }
        },
        "summary": "Revoke a token",
        "tags": [
          "invalid-tokens"
        ]
      }
    },
    "/invalid-tokens/{id}": {
      "get": {
        "operationId": "InvalidTokensController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Denylist entry ID",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Denylist entry found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvalidTokenResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Denylist entry not found."
          }
        },
        "summary": "Get a single denylist entry",
        "tags": [
          "invalid-tokens"
        ]
      },
      "patch": {
        "operationId": "InvalidTokensController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID of the entry to update",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "description": "New data for the denylist entry.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateInvalidTokenDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Denylist entry updated successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvalidTokenCreatedResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid data — the jti is already on the denylist."
          },
          "404": {
            "description": "Denylist entry not found."
          }
        },
        "summary": "Update a denylist entry",
        "tags": [
          "invalid-tokens"
        ]
      },
      "delete": {
        "operationId": "InvalidTokensController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Denylist entry ID",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Denylist entry deleted successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Denylist entry not found."
          }
        },
        "summary": "Delete a denylist entry by its ID",
        "tags": [
          "invalid-tokens"
        ]
      }
    },
...
  },
  "info": {
    "title": "Spira API Documentation",
    "description": "Spira API",
    "version": "0.0.1",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "MessageResponseDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "Human-readable message describing the outcome.",
            "example": "Request completed successfully"
          }
        },
        "required": [
          "message"
        ]
      },
...
      "ContactType": {
        "type": "string",
        "enum": [
          "PRIMARY",
          "OPERATIONS",
          "LOGISTICS",
          "BILLING",
          "EMERGENCY"
        ],
        "description": "Which function this contact covers."
      },
      "ContactResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique contact ID.",
            "format": "uuid"
          },
          "retailerId": {
            "type": "object",
            "description": "Owning retailer. Exactly one owner is set.",
            "format": "uuid",
            "nullable": true
          },
          "recipientId": {
            "type": "object",
            "description": "Owning recipient. Exactly one owner is set.",
            "format": "uuid",
            "nullable": true
          },
          "fullName": {
            "type": "string",
            "description": "Full name of the person.",
            "example": "María González"
          },
          "email": {
            "type": "object",
            "description": "Email address.",
            "nullable": true
          },
          "phone": {
            "type": "object",
            "description": "Phone number in E.164 format.",
            "nullable": true
          },
          "secondaryPhone": {
            "type": "object",
            "description": "Alternate phone number in E.164 format.",
            "nullable": true
          },
          "jobTitle": {
            "type": "object",
            "description": "Role inside the organization.",
            "nullable": true
          },
          "type": {
            "description": "Which function this contact covers.",
            "example": "OPERATIONS",
            "allOf": [
              {
                "$ref": "#/components/schemas/ContactType"
              }
            ]
          },
          "isPrimary": {
            "type": "boolean",
            "description": "Whether this is the main contact."
          },
          "notes": {
            "type": "object",
            "description": "Internal notes.",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "description": "When the contact was created (ISO 8601).",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "description": "When the contact was last modified (ISO 8601).",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "fullName",
          "type",
          "isPrimary",
          "createdAt",
          "updatedAt"
        ]
      },
      "CreateContactDto": {
        "type": "object",
        "properties": {
          "retailerId": {
            "type": "string",
            "description": "Owning retailer. Provide exactly one of `retailerId` / `recipientId`.",
            "format": "uuid",
            "example": "3f2c1b8e-9a4d-4c7f-8b1e-2d6a5c9f0e11"
          },
          "recipientId": {
            "type": "string",
            "description": "Owning recipient. Provide exactly one of `retailerId` / `recipientId`.",
            "format": "uuid",
            "example": "7a1e4c9d-2b8f-4e6a-9c30-5d7b1f2a8c44"
          },
          "fullName": {
            "type": "string",
            "description": "Full name of the person.",
            "example": "María González",
            "minLength": 2,
            "maxLength": 150
          },
          "email": {
            "type": "string",
            "description": "Email address. Stored case-insensitively.",
            "example": "maria.gonzalez@real.com.py"
          },
          "phone": {
            "type": "string",
            "description": "Phone number in E.164 format.",
            "example": "+595981123456",
            "maxLength": 30
          },
          "secondaryPhone": {
            "type": "string",
            "description": "Alternate phone number in E.164 format.",
            "example": "+595981654321",
            "maxLength": 30
          },
          "jobTitle": {
            "type": "string",
            "description": "Role inside the organization.",
            "example": "Head of Operations",
            "maxLength": 100
          },
          "type": {
            "description": "Which function this contact covers.",
            "example": "OPERATIONS",
            "allOf": [
              {
                "$ref": "#/components/schemas/ContactType"
              }
            ]
          },
          "isPrimary": {
            "type": "boolean",
            "description": "Marks the main contact for the owner. At most one per owner.",
            "default": false,
            "example": false
          },
          "notes": {
            "type": "string",
            "description": "Internal notes about reaching this person.",
            "example": "Prefers WhatsApp. Unavailable Mondays."
          }
        },
        "required": [
          "fullName",
          "type"
        ]
      },
      "ContactCreatedResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique contact ID.",
            "format": "uuid"
          },
          "retailerId": {
            "type": "object",
            "description": "Owning retailer. Exactly one owner is set.",
            "format": "uuid",
            "nullable": true
          },
          "recipientId": {
            "type": "object",
            "description": "Owning recipient. Exactly one owner is set.",
            "format": "uuid",
            "nullable": true
          },
          "fullName": {
            "type": "string",
            "description": "Full name of the person.",
            "example": "María González"
          },
          "email": {
            "type": "object",
            "description": "Email address.",
            "nullable": true
          },
          "phone": {
            "type": "object",
            "description": "Phone number in E.164 format.",
            "nullable": true
          },
          "secondaryPhone": {
            "type": "object",
            "description": "Alternate phone number in E.164 format.",
            "nullable": true
          },
          "jobTitle": {
            "type": "object",
            "description": "Role inside the organization.",
            "nullable": true
          },
          "type": {
            "description": "Which function this contact covers.",
            "example": "OPERATIONS",
            "allOf": [
              {
                "$ref": "#/components/schemas/ContactType"
              }
            ]
          },
          "isPrimary": {
            "type": "boolean",
            "description": "Whether this is the main contact."
          },
          "notes": {
            "type": "object",
            "description": "Internal notes.",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "description": "When the contact was created (ISO 8601).",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "description": "When the contact was last modified (ISO 8601).",
            "format": "date-time"
          },
          "message": {
            "type": "string",
            "description": "Human-readable message describing the outcome.",
            "example": "Contact created successfully."
          }
        },
        "required": [
          "id",
          "fullName",
          "type",
          "isPrimary",
          "createdAt",
          "updatedAt",
          "message"
        ]
      },
      "UpdateContactDto": {
        "type": "object",
        "properties": {
          "retailerId": {
            "type": "string",
            "description": "Owning retailer. Provide exactly one of `retailerId` / `recipientId`.",
            "format": "uuid",
            "example": "3f2c1b8e-9a4d-4c7f-8b1e-2d6a5c9f0e11"
          },
          "recipientId": {
            "type": "string",
            "description": "Owning recipient. Provide exactly one of `retailerId` / `recipientId`.",
            "format": "uuid",
            "example": "7a1e4c9d-2b8f-4e6a-9c30-5d7b1f2a8c44"
          },
          "fullName": {
            "type": "string",
            "description": "Full name of the person.",
            "example": "María González",
            "minLength": 2,
            "maxLength": 150
          },
          "email": {
            "type": "string",
            "description": "Email address. Stored case-insensitively.",
            "example": "maria.gonzalez@real.com.py"
          },
          "phone": {
            "type": "string",
            "description": "Phone number in E.164 format.",
            "example": "+595981123456",
            "maxLength": 30
          },
          "secondaryPhone": {
            "type": "string",
            "description": "Alternate phone number in E.164 format.",
            "example": "+595981654321",
            "maxLength": 30
          },
          "jobTitle": {
            "type": "string",
            "description": "Role inside the organization.",
            "example": "Head of Operations",
            "maxLength": 100
          },
          "type": {
            "description": "Which function this contact covers.",
            "example": "OPERATIONS",
            "allOf": [
              {
                "$ref": "#/components/schemas/ContactType"
              }
            ]
          },
          "isPrimary": {
            "type": "boolean",
            "description": "Marks the main contact for the owner. At most one per owner.",
            "default": false,
            "example": false
          },
          "notes": {
            "type": "string",
            "description": "Internal notes about reaching this person.",
            "example": "Prefers WhatsApp. Unavailable Mondays."
          }
        }
      },
...
      "InvalidTokenReason": {
        "type": "string",
        "enum": [
          "LOGOUT",
          "PASSWORD_CHANGE",
          "ADMIN_REVOKE",
          "SECURITY"
        ],
        "description": "Why the token was revoked."
      },
      "InvalidTokenResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique denylist entry ID.",
            "format": "uuid"
          },
          "jti": {
            "type": "string",
            "description": "The revoked token's `jti` claim.",
            "format": "uuid"
          },
          "userId": {
            "type": "string",
            "description": "Owner of the revoked token.",
            "format": "uuid"
          },
          "expiresAt": {
            "type": "string",
            "description": "When the token naturally expires (ISO 8601).",
            "format": "date-time"
          },
          "reason": {
            "description": "Why the token was revoked.",
            "example": "LOGOUT",
            "allOf": [
              {
                "$ref": "#/components/schemas/InvalidTokenReason"
              }
            ]
          },
          "isExpired": {
            "type": "boolean",
            "description": "Whether the token is already past its expiry.",
            "example": false
          },
          "createdAt": {
            "type": "string",
            "description": "When the token was invalidated (ISO 8601).",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "description": "When the entry was last modified (ISO 8601).",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "jti",
          "userId",
          "expiresAt",
          "reason",
          "isExpired",
          "createdAt",
          "updatedAt"
        ]
      },
      "CreateInvalidTokenDto": {
        "type": "object",
        "properties": {
          "jti": {
            "type": "string",
            "description": "The revoked token's `jti` claim. Must be unique.",
            "format": "uuid",
            "example": "9b1d4f2a-6c3e-4a8b-9f10-7e5c2d4a8b16"
          },
          "userId": {
            "type": "string",
            "description": "Owner of the revoked token.",
            "format": "uuid",
            "example": "3f2c1b8e-9a4d-4c7f-8b1e-2d6a5c9f0e11"
          },
          "expiresAt": {
            "type": "string",
            "description": "When the token naturally expires (the JWT `exp`).",
            "format": "date-time",
            "example": "2026-09-16T15:32:05.000Z"
          },
          "reason": {
            "description": "Why the token was revoked.",
            "example": "LOGOUT",
            "allOf": [
              {
                "$ref": "#/components/schemas/InvalidTokenReason"
              }
            ]
          }
        },
        "required": [
          "jti",
          "userId",
          "expiresAt",
          "reason"
        ]
      },
      "InvalidTokenCreatedResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique denylist entry ID.",
            "format": "uuid"
          },
          "jti": {
            "type": "string",
            "description": "The revoked token's `jti` claim.",
            "format": "uuid"
          },
          "userId": {
            "type": "string",
            "description": "Owner of the revoked token.",
            "format": "uuid"
          },
          "expiresAt": {
            "type": "string",
            "description": "When the token naturally expires (ISO 8601).",
            "format": "date-time"
          },
          "reason": {
            "description": "Why the token was revoked.",
            "example": "LOGOUT",
            "allOf": [
              {
                "$ref": "#/components/schemas/InvalidTokenReason"
              }
            ]
          },
          "isExpired": {
            "type": "boolean",
            "description": "Whether the token is already past its expiry.",
            "example": false
          },
          "createdAt": {
            "type": "string",
            "description": "When the token was invalidated (ISO 8601).",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "description": "When the entry was last modified (ISO 8601).",
            "format": "date-time"
          },
          "message": {
            "type": "string",
            "description": "Human-readable message describing the outcome.",
            "example": "Token revoked successfully."
          }
        },
        "required": [
          "id",
          "jti",
          "userId",
          "expiresAt",
          "reason",
          "isExpired",
          "createdAt",
          "updatedAt",
          "message"
        ]
      },
...
    }
  }
}